+++ src/main/java/courgette/runtime/CourgetteHtmlReporter.java
@@ -2,9 +2,7 @@ package courgette.runtime;
 
 import courgette.api.CourgetteRunLevel;
 import courgette.runtime.report.builder.HtmlReportBuilder;
-import courgette.runtime.report.model.Embedding;
 import courgette.runtime.report.model.Feature;
-import courgette.runtime.report.model.Step;
 import courgette.runtime.utils.FileUtils;
 
 import java.io.BufferedReader;
@@ -16,18 +14,11 @@ import java.time.temporal.ChronoUnit;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
-import java.util.function.Consumer;
 
 public class CourgetteHtmlReporter {
     private final String INDEX_HTML = "/report/index.html";
-    private final String CSS_ASSETS = "/report/css/bootstrap.min.css,/report/css/core.min.css,/report/css/dataTables.bootstrap4.css";
-    private final String JS_ASSETS = "/report/js/bootstrap.min.js,/report/js/core.min.js,/report/js/dataTables.bootstrap4.js,/report/js/jquery.dataTables.js,/report/js/jquery.easing.min.js,/report/js/jquery.min.js,/report/js/popper.min.js,/report/js/mdb.min.js,/report/js/Chart.min.js";
-
     private final String targetDir;
     private final String reportDir;
-    private final String imagesDir;
-    private final String cssDir;
-    private final String jsDir;
 
     private final CourgetteProperties courgetteProperties;
     private final List<CourgetteRunResult> courgetteRunResults;
@@ -36,10 +27,6 @@ public class CourgetteHtmlReporter {
     public CourgetteHtmlReporter(CourgetteProperties courgetteProperties, List<CourgetteRunResult> courgetteRunResults, List<Feature> reportFeatures) {
         this.targetDir = courgetteProperties.getCourgetteOptions().reportTargetDir();
         this.reportDir = targetDir + "/courgette-report";
-        this.imagesDir = reportDir + "/images";
-        this.cssDir = reportDir + "/css";
-        this.jsDir = reportDir + "/js";
-
         this.courgetteProperties = courgetteProperties;
         this.courgetteRunResults = courgetteRunResults;
         this.reportFeatures = reportFeatures;
@@ -48,7 +35,6 @@ public class CourgetteHtmlReporter {
     public void create() {
         createReportDirectories();
         generateHtmlReport();
-        copyReportAssets();
     }
 
     private void generateHtmlReport() {
@@ -104,26 +90,6 @@ public class CourgetteHtmlReporter {
 
         formattedIndexHtml = formattedIndexHtml.replaceAll("id:features", cucumberFeatures);
 
-        final Consumer<Embedding> saveImage = (embedding) -> {
-            if (embedding.getMimeType().startsWith("image")) {
-                final String imageName = imagesDir + "/" + embedding.getCourgetteEmbeddingId();
-                final String imageFormat = embedding.getMimeType().split("/")[1];
-                FileUtils.writeImageFile(imageName, imageFormat, embedding.getData());
-            }
-        };
-
-        reportFeatures.forEach(feature ->
-                feature.getScenarios().forEach(scenario -> {
-                    scenario.getBefore().forEach(before -> before.getEmbeddings().forEach(saveImage));
-                    scenario.getAfter().forEach(after -> after.getEmbeddings().forEach(saveImage));
-
-                    List<Step> scenarioSteps = scenario.getSteps();
-
-                    scenarioSteps.forEach(step -> step.getEmbeddings().forEach(saveImage));
-                    scenarioSteps.forEach(step -> step.getBefore().forEach(before -> before.getEmbeddings().forEach(saveImage)));
-                    scenarioSteps.forEach(step -> step.getAfter().forEach(after -> after.getEmbeddings().forEach(saveImage)));
-                }));
-
         final HtmlReportBuilder htmlReportBuilder = HtmlReportBuilder.create(reportFeatures, courgetteRunResults, isStrict);
 
         final String results = courgetteProperties.getCourgetteOptions().runLevel() == CourgetteRunLevel.FEATURE ?
@@ -134,22 +100,6 @@ public class CourgetteHtmlReporter {
         formattedIndexHtml = formattedIndexHtml.replace("id:modals", htmlReportBuilder.getHtmlModals());
 
         FileUtils.writeFile(reportDir + "/index.html", formattedIndexHtml);
-    }
-
-    private void copyReportAssets() {
-        Arrays.stream(CSS_ASSETS.split(",")).forEach(cssAsset -> {
-            InputStream resource = getClass().getResourceAsStream(cssAsset);
-            if (resource != null) {
-                FileUtils.readAndWriteFile(resource, cssDir + "/" + cssAsset.substring(cssAsset.lastIndexOf("/") + 1));
-            }
-        });
-
-        Arrays.stream(JS_ASSETS.split(",")).forEach(jsAsset -> {
-            InputStream resource = getClass().getResourceAsStream(jsAsset);
-            if (resource != null) {
-                FileUtils.readAndWriteFile(resource, jsDir + "/" + jsAsset.substring(jsAsset.lastIndexOf("/") + 1));
-            }
-        });
     }
 
     private void createReportDirectories() {
@@ -166,30 +116,6 @@ public class CourgetteHtmlReporter {
         if (!reportDir.exists()) {
             if (!reportDir.mkdir()) {
                 throw new CourgetteException("Unable to create the '../courgette-report' directory");
-            }
-        }
-
-        final File cssDir = new File(this.cssDir);
-
-        if (!cssDir.exists()) {
-            if (!cssDir.mkdir()) {
-                throw new CourgetteException("Unable to create the '../css' directory");
-            }
-        }
-
-        final File jsDir = new File(this.jsDir);
-
-        if (!jsDir.exists()) {
-            if (!jsDir.mkdir()) {
-                throw new CourgetteException("Unable to create the '../js' directory");
-            }
-        }
-
-        final File imagesDir = new File(this.imagesDir);
-
-        if (!imagesDir.exists()) {
-            if (!imagesDir.mkdir()) {
-                throw new CourgetteException("Unable to create the '../images' directory");
             }
         }
     }